set user name on index page
