+++ components/NavigationLink/index.js
@@ -7,7 +7,7 @@ const StyledNavigationLink = styled.a`
   border-radius: 5px;
   padding: 5px;
   margin: 5px;
-  font-size: 24px;
+  font-size: 20px;
   text-decoration: none;
   background-color: ${(props) =>
     props.selected ? "hsla(223, 57%, 77%, 1)" : "#d96a3f"};
@@ -17,7 +17,7 @@ const StyledNavigationLink = styled.a`
   }
 `;
 
-export default function NavigationLink({ children, href, selected }) {
+export default function NavigationLink({children, href, selected}) {
   return (
     <StyledNavigationLink selected={selected} href={href}>
       {children}
