+++ pages/_app.js
@@ -1,20 +1,27 @@
 import GlobalStyle from "../styles";
 import Head from "next/head";
-import { useState } from "react";
+import {useState} from "react";
 import useLocalStorageState from "use-local-storage-state";
 
-export default function App({ Component, pageProps }) {
+export default function App({Component, pageProps}) {
   const [userName, setUserName] = useLocalStorageState("fitnessAppUserName", {
     defalutValue: "",
   });
-
+  function handleUserNameFormSubmit(event, userInput) {
+    event.preventDefault();
+    setUserName(userInput);
+  }
   return (
     <>
       <GlobalStyle />
       <Head>
-        <title>Capstone Project</title>
+        <title>Capstone Project Fitness App</title>
       </Head>
-      <Component userName={userName} setUserName={setUserName} {...pageProps} />
+      <Component
+        userName={userName}
+        handleUserNameFormSubmit={handleUserNameFormSubmit}
+        {...pageProps}
+      />
     </>
   );
 }
