+++ pages/index.js
@@ -1,12 +1,20 @@
 import Heading from "../components/Heading";
-import Link from "next/link";
-
-export default function Home() {
+import UserNameForm from "../components/UserNameForm";
+import Header from "../components/Header";
+import Navigation from "../components/Navigation";
+export default function Home({userName, handleUserNameFormSubmit}) {
   return (
-    <main>
-      <Heading>🔱Fitness App🔱</Heading>
-      <Link href="/calendar">Calendar</Link>
-      <Link href="/strength">Strength</Link>
-    </main>
+    <>
+      <main>
+        <Header>
+          <Heading>Fitness App</Heading>
+        </Header>
+        <UserNameForm
+          userName={userName}
+          handleUserNameFormSubmit={handleUserNameFormSubmit}
+        ></UserNameForm>
+      </main>
+      <Navigation selected={"index"} userName={userName}></Navigation>
+    </>
   );
 }
